rename utils.js to meshUtils.js and add to new utils directory
